Extract nvim state manager out of settings struct and rename Settings to SettingsManager
