+++ src/renderer/rendered_window.rs
@@ -97,7 +97,7 @@ pub struct WindowDrawDetails {
 
 impl WindowDrawDetails {
     pub fn event_grid_id(&self) -> u64 {
-        if SETTINGS.get::<CmdLineSettings>().no_multi_grid {
+        if SETTINGS.get::<CmdLineSettings>().no_multigrid {
             0
         } else {
             self.id
